Add features

+++ src/main/java/com/inversionesaraujo/api/repository/AdvanceRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.Advance;
 
-public interface AdvanceRepository extends JpaRepository<Advance, Integer> {}
+public interface AdvanceRepository extends JpaRepository<Advance, Long> {}
