+++ src/main/java/com/inversionesaraujo/api/repository/ImageRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.Image;
 
-public interface ImageRepository extends JpaRepository<Image, Integer> {}
+public interface ImageRepository extends JpaRepository<Image, Long> {}
