+++ src/main/java/com/inversionesaraujo/api/repository/OrderProductRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.OrderProduct;
 
-public interface OrderProductRepository extends JpaRepository<OrderProduct, Integer> {}
+public interface OrderProductRepository extends JpaRepository<OrderProduct, Long> {}
