+++ src/main/java/com/inversionesaraujo/api/repository/ProfitRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.Profit;
 
-public interface ProfitRepository extends JpaRepository<Profit, Integer> {
+public interface ProfitRepository extends JpaRepository<Profit, Long> {
     Profit findByMonth(String month);
 }
